backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "REST API онлайн выставки котят"
authors = ["Екатерина Хиталенко"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
djoser = "^2.2.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
